+++ integration/ruler_test.go
@@ -97,7 +97,7 @@ func TestRulerAPI(t *testing.T) {
 	require.Equal(t, retrievedNamespace[0].Name, ruleGroup.Name)
 
 	// Test compression by inspecting the response Headers
-	req, err := http.NewRequest("GET", fmt.Sprintf("http://%s/api/prom/rules", ruler.HTTPEndpoint()), nil)
+	req, err := http.NewRequest("GET", fmt.Sprintf("http://%s/api/v1/rules", ruler.HTTPEndpoint()), nil)
 	require.NoError(t, err)
 
 	req.Header.Set("X-Scope-OrgID", "user-1")
@@ -830,14 +830,19 @@ func TestRulerEnableAPIs(t *testing.T) {
 			apiEnabled: false,
 
 			expectedRegisteredEndpoints: [][2]string{
-				{http.MethodGet, "/api/prom/api/v1/alerts"},
-				{http.MethodGet, "/api/prom/api/v1/rules"},
+				{http.MethodGet, "/prometheus/api/v1/alerts"},
+				{http.MethodGet, "/prometheus/api/v1/rules"},
 			},
 			expectedMissingEndpoints: [][2]string{
 				{http.MethodGet, "/api/v1/rules"},
 				{http.MethodGet, "/api/v1/rules/my_namespace"},
 				{http.MethodGet, "/api/v1/rules/my_namespace/my_group"},
 				{http.MethodPost, "/api/v1/rules/my_namespace"},
+
+				{http.MethodGet, "/prometheus/rules"},
+				{http.MethodGet, "/prometheus/rules/my_namespace"},
+				{http.MethodGet, "/prometheus/rules/my_namespace/my_group"},
+				{http.MethodPost, "/prometheus/rules/my_namespace"},
 			},
 		},
 		{
@@ -846,11 +851,16 @@ func TestRulerEnableAPIs(t *testing.T) {
 
 			expectedRegisteredEndpoints: [][2]string{
 				// not going to test GET /api/v1/rules/my_namespace/my_group because it requires creating a rule group
-				{http.MethodGet, "/api/prom/api/v1/alerts"},
-				{http.MethodGet, "/api/prom/api/v1/rules"},
+				{http.MethodGet, "/prometheus/api/v1/alerts"},
+				{http.MethodGet, "/prometheus/api/v1/rules"},
+
 				{http.MethodGet, "/api/v1/rules"},
 				{http.MethodGet, "/api/v1/rules/my_namespace"},
 				{http.MethodPost, "/api/v1/rules/my_namespace"},
+
+				{http.MethodGet, "/prometheus/rules"},
+				{http.MethodGet, "/prometheus/rules/my_namespace"},
+				{http.MethodPost, "/prometheus/rules/my_namespace"},
 			},
 		},
 	}