+++ pkg/ingester/ingester_test.go
@@ -3408,7 +3408,7 @@ func TestIngester_flushing(t *testing.T) {
 		cortex_ingester_shipper_uploads_total 0
 	`), "cortex_ingester_shipper_uploads_total"))
 
-				i.ShutdownHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/shutdown", nil))
+				i.ShutdownHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/ingester/shutdown", nil))
 
 				verifyCompactedHead(t, i, true)
 				require.NoError(t, testutil.GatherAndCompare(reg, bytes.NewBufferString(`
@@ -3435,7 +3435,7 @@ func TestIngester_flushing(t *testing.T) {
 				`), "cortex_ingester_shipper_uploads_total"))
 
 				// Using wait=true makes this a synchronous call.
-				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/flush?wait=true", nil))
+				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/ingester/flush?wait=true", nil))
 
 				verifyCompactedHead(t, i, true)
 				require.NoError(t, testutil.GatherAndCompare(reg, bytes.NewBufferString(`
@@ -3466,7 +3466,7 @@ func TestIngester_flushing(t *testing.T) {
 				users.Add(tenantParam, "another-unknown-user")
 
 				// Using wait=true makes this a synchronous call.
-				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/flush?wait=true&"+users.Encode(), nil))
+				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/ingester/flush?wait=true&"+users.Encode(), nil))
 
 				// Still nothing shipped or compacted.
 				require.NoError(t, testutil.GatherAndCompare(reg, bytes.NewBufferString(`
@@ -3481,7 +3481,7 @@ func TestIngester_flushing(t *testing.T) {
 				users.Add(tenantParam, userID) // Our user
 				users.Add(tenantParam, "yet-another-user")
 
-				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/flush?wait=true&"+users.Encode(), nil))
+				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/ingester/flush?wait=true&"+users.Encode(), nil))
 
 				verifyCompactedHead(t, i, true)
 				require.NoError(t, testutil.GatherAndCompare(reg, bytes.NewBufferString(`
@@ -3517,7 +3517,7 @@ func TestIngester_flushing(t *testing.T) {
 					cortex_ingester_shipper_uploads_total 0
 				`), "cortex_ingester_shipper_uploads_total"))
 
-				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/flush?wait=true", nil))
+				i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/ingester/flush?wait=true", nil))
 
 				verifyCompactedHead(t, i, true)
 
@@ -4247,7 +4247,7 @@ func TestIngesterNoFlushWithInFlightRequest(t *testing.T) {
 
 	// Flush handler only triggers compactions, but doesn't wait for them to finish. We cannot use ?wait=true here,
 	// because it would deadlock -- flush will wait for appendLock to be released.
-	i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/flush", nil))
+	i.FlushHandler(httptest.NewRecorder(), httptest.NewRequest("POST", "/ingester/flush", nil))
 
 	// Flushing should not have succeeded even after 5 seconds.
 	time.Sleep(5 * time.Second)