+++ pkg/mimir/modules.go
@@ -95,7 +95,6 @@ func newDefaultConfig() *Config {
 
 func (t *Mimir) initAPI() (services.Service, error) {
 	t.Cfg.API.ServerPrefix = t.Cfg.Server.PathPrefix
-	t.Cfg.API.LegacyHTTPPrefix = t.Cfg.HTTPPrefix
 
 	a, err := api.New(t.Cfg.API, t.Cfg.Server, t.Server, util_log.Logger)
 	if err != nil {
@@ -325,11 +324,11 @@ func (t *Mimir) initTenantFederation() (serv services.Service, err error) {
 //        requests     │                     │                     │
 //                     │                     │                     │
 //                     ▼                     ▼                     ▼
-//           ┌──────────────────┐  ┌──────────────────┐  ┌──────────────────┐
-//           │                  │  │                  │  │                  │
-//           │Querier Queryable │  │  /api/v1 router  │  │ /api/prom router │
-//           │                  │  │                  │  │                  │
-//           └──────────────────┘  └──────────────────┘  └──────────────────┘
+//           ┌──────────────────┐  ┌──────────────────┐  ┌────────────────────┐
+//           │                  │  │                  │  │                    │
+//           │Querier Queryable │  │  /api/v1 router  │  │ /prometheus router │
+//           │                  │  │                  │  │                    │
+//           └──────────────────┘  └──────────────────┘  └────────────────────┘
 //                     ▲                     │                     │
 //                     │                     └──────────┬──────────┘
 //                     │                                ▼
@@ -603,7 +602,7 @@ func (t *Mimir) initAlertManager() (serv services.Service, err error) {
 		return
 	}
 
-	t.API.RegisterAlertmanager(t.Alertmanager, t.Cfg.isModuleEnabled(AlertManager), t.Cfg.Alertmanager.EnableAPI)
+	t.API.RegisterAlertmanager(t.Alertmanager, t.Cfg.Alertmanager.EnableAPI)
 	return t.Alertmanager, nil
 }